move MySQLAuthenticationHandler from sharding-proxy-transport-mysql to sharding-proxy-frontend-mysql
